requirements.txt
aws-cdk.aws-s3
aws_cdk.aws_route53
aws_cdk.aws_route53_targets